Extracts tests from Listener spec to RequestHandlers::Message spec file; adds test to Listener spec file to ensure that RequestHandlers::Message is called with the correct arguments.
